Reorganize files: pull service models into lib folder
