+++ lib/prompt_input_strategy.rb
@@ -1,0 +1,5 @@
+class PromptInputStrategy
+  def prompt(value)
+    $prompt.ask(value)
+  end
+end
