+++ main.rb
@@ -1,3 +1,4 @@
-require_relative './models/interface'
+require_relative './lib/interface'
+require_relative './lib/prompt_input_strategy'
 
 Interface.new(PromptInputStrategy.new).start
